refactor: move controllers into controllers folder
